fabpricetracker/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.13
django-bootstrap-v5==1.0.11
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
exceptiongroup==1.2.1
flower==2.0.1
h11==0.14.0
humanize==4.9.0
idna==3.7
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
outcome==1.3.0.post0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
pyparsing==3.1.2
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.21.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stealth==0.0.1
tornado==6.4.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websockets==13.0.1
wsproto==1.2.0
